# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pakperk API",
    "description": "Stable v1 paper-reading API with optional OIDC-backed account profiles. Library and comment routes remain feature-gated until their production phases are complete.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/papers/by-arxiv/{arxiv_id}": {
      "get": {
        "tags": [
          "routes::papers"
        ],
        "operationId": "paper_by_arxiv",
        "parameters": [
          {
            "name": "arxiv_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paper metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSummarySchema"
                }
              }
            }
          },
          "400": {
            "description": "Invalid arXiv identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Paper not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
    "/v1/papers/{paper_id}": {
      "get": {
        "tags": [
          "routes::papers"
        ],
        "operationId": "paper_metadata",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paper metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaperSummarySchema"
                }
              }
            }
          },
          "404": {
            "description": "Paper not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
    "/v1/papers/{paper_id}/chat": {
      "post": {
        "tags": [
          "routes::chat"
        ],
        "operationId": "chat",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Grounded chat answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseSchema"
                }
              }
            }
          },
          "400": {
            "description": "Invalid anonymous session or question",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "403": {
            "description": "Full-text policy denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "409": {
            "description": "Capability not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
    "/v1/papers/{paper_id}/connections": {
      "get": {
        "tags": [
          "routes::papers"
        ],
        "operationId": "connections",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paper connections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionsResponseSchema"
                }
              }
            }
          },
          "403": {
            "description": "Full-text policy denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Paper not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "409": {
            "description": "Capability not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
    "/v1/papers/{paper_id}/introduction": {
      "get": {
        "tags": [
          "routes::papers"
        ],
        "operationId": "introduction",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paper introduction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntroductionSchema"
                }
              }
            }
          },
          "403": {
            "description": "Full-text policy denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Paper not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "409": {
            "description": "Capability not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
    "/v1/papers/{paper_id}/prepare": {
      "post": {
        "tags": [
          "routes::papers"
        ],
        "operationId": "prepare",
        "parameters": [
          {
            "name": "paper_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Already ready or terminal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingStateSchema"
                }
              }
            }
          },
          "202": {
            "description": "Preparation accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessingStateSchema"
                }
              }
            }
          },
          "404": {
            "description": "Paper not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeSchema"
                }
              }
            }
          }
        },
        "security": [
          {},
          {
            "oidcBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorEnvelopeSchema": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBodySchema"
          }
        }
      },
...
      "PaperSummarySchema": {
        "type": "object",
        "required": [
          "paper_id",
          "arxiv_id",
          "title",
          "abstract",
          "authors",
          "primary_category",
          "categories",
          "published_at",
          "updated_at",
          "abs_url",
          "pdf_url",
          "capabilities"
        ],
        "properties": {
          "abs_url": {
            "type": "string"
          },
          "abstract": {
            "type": "string"
          },
          "arxiv_id": {
            "type": "string",
            "example": "2401.12345v2"
          },
          "authors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "capabilities": {
            "$ref": "#/components/schemas/CapabilitiesSchema"
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "paper_id": {
            "type": "string",
            "format": "uuid"
          },
          "pdf_url": {
            "type": "string"
          },
          "primary_category": {
            "type": "string"
          },
          "published_at": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "oidcBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OIDC access token validated against the configured issuer, audience, algorithm, and JWKS."
      }
    }
  },
...
}